% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_OnlineSelection.R
\name{CiterionCompute_each}
\alias{CiterionCompute_each}
\title{function to calculate FSP and ES given the decisions and the truth for data until time t}
\usage{
CiterionCompute_each(X_test_scale, Alter_test, decisions, t, x.Real)
}
\arguments{
\item{X_test_scale:}{The center scaled covariates X}

\item{Alter_test:}{the index of true signal  (theta=1)}

\item{decisions:}{the vector of decisions}

\item{t:}{time point t before stopping}

\item{x.Real:}{the true theta until the stopping time}
}
\value{
the FSP (FDP) and ES (DIV) at every time t
}
\description{
function to calculate FSP and ES given the decisions and the truth for data until time t
}
