% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_OnlineSelection.R
\name{Online_selection_correct_rej}
\alias{Online_selection_correct_rej}
\title{function to carry out the DOSS online sample selection procedure}
\usage{
Online_selection_correct_rej(
  X_test_scale,
  Alter_test,
  TN,
  alpha,
  Diversity_threshold,
  Diversity_initial_num,
  m,
  N,
  IsDiversity
)
}
\arguments{
\item{X_test_scale}{The center scaled covariates X}

\item{Alter_test}{the index of true signal (theta=1)}

\item{TN}{The local FDR}

\item{alpha}{The FSR level alpha}

\item{Diversity_threshold}{The ES level K}

\item{Diversity_initial_num}{When the selection number exceeds the Diversity_initial_num, we consider computing diversity.}

\item{m}{the number of selection when stop}

\item{N}{the total number of observations}

\item{IsDiversity}{A binary variable. If IsDiversity==TRUE, the DOSS procedure is carried out, else the SAST procedure is carried out.}
}
\value{
A list contains FSP (FDP), ES (DIV), Power, decisions, select.num, and stoptime.
}
\description{
function to carry out the DOSS online sample selection procedure
}
